+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Breaking: Remove move ctor/operator for SettingListener. (#32)
+- Dev: Remove `using namespace std` usages. (#38)
 
 ## v0.1.0